# docs/openapi/quiket-mvp-api.yaml
openapi: 3.0.3
info:
  title: Quiket MVP API
  version: 1.0.1
  description: |
    Quiket MVP API 명세

    기준 문서:
    - 기존 Quiket Auth API 명세
    - Quiket 기능명세 목록 v1.0
    - Quiket 기능명세 상세 v1.0
    - Quiket 서비스 DDL

    MVP 범위:
    - 자체 회원가입, 이메일 인증, 자체 로그인
    - JWT Access Token / Refresh Token 재발급, 로그아웃
    - 비밀번호 재설정
    - Kakao OAuth 로그인 및 기존 계정 연동
    - 홈 메인 데이터, 최근활동
    - 과목 추가/조회/수정/삭제, 시험 D-Day 설정
    - 챕터명 수정
    - PDF/이미지/텍스트 강의 업로드
    - 파트 자동/직접 분류, 파트 조회/편집/추가
    - 퀴즈 생성, 생성 상태 조회, 퀴즈 세트 조회
    - 퀴즈 결과 제출, 결과/해설 조회, 다시 풀기
    - 도토리/XP/캐릭터 성장 조회
    - 마이페이지 계정(닉네임/이메일 변경, 비밀번호 변경, 회원 탈퇴), 알림, 피드백

    MVP 제외:
    - 퀴즈기록 별도 메뉴
    - 오답노트 별도 메뉴
    - PPTX/DOCX 업로드
    - 텍스트 검수 화면
    - 파트 병합/분리/순서변경/삭제
    - 플래시카드, 쪽지시험
    - 아이템 스토어
    - 요점정리

    공통 응답 형식:
    - 성공/실패 응답은 모두 ApiResponse 형태를 사용합니다.
    - success 필드는 성공/실패 여부만 표현합니다.
    - 외부 노출 식별자는 UUID v7 문자열(public_id)을 사용합니다.

    [Redis 사용 정책]
    - 이메일 인증 코드, 비밀번호 재설정 코드, Kakao signupToken, Kakao linkToken은 Redis TTL 10분 기반 임시 데이터로 관리합니다.
    - 이메일 인증 코드와 비밀번호 재설정 코드는 재발송 시 이전 pending 코드를 즉시 무효화합니다.
    - Refresh Token은 DB에 해시값으로 영속 저장합니다.
    - Redis 임시 데이터는 만료 후 재사용하지 않습니다.
servers:
  - url: http://43.201.222.243:8080/api/v1
    description: Quiket 운영/개발 서버
  - url: http://localhost:8080/api/v1
    description: Local server
...
paths:
...
  /home:
    get:
      tags:
        - Home
      summary: 홈 메인 데이터 조회
      description: |
        홈 메인 화면에 필요한 사용자 요약, 캐릭터 바, D-Day, 과목 목록, 최근활동을 한 번에 조회합니다.
        기록/오답 별도 메뉴는 MVP에서 제외되며 최근활동은 홈에서만 노출합니다.
      operationId: getHome
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 홈 데이터 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /home/recent-activities:
    get:
      tags:
        - Home
        - History
      summary: 홈 최근활동 목록 조회
      description: 퀴즈 생성 중, 미시작, 진행 중, 완료 상태의 최근활동을 최신순으로 조회합니다.
      operationId: getRecentActivities
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 최근활동 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecentActivityListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      required: false
      description: 0부터 시작하는 페이지 번호
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      description: 페이지 크기
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
  responses:
...
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - code
        - message
        - data
      properties:
        success:
          type: boolean
          example: false
        code:
          type: string
          example: COMMON_VALIDATION_ERROR
        message:
          type: string
          example: 요청 값이 올바르지 않습니다.
        data:
          type: object
          nullable: true
          additionalProperties: true
          example: null
...
    HomeResponse:
      allOf:
        - $ref: '#/components/schemas/ApiSuccessBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/HomeData'
    HomeData:
      type: object
      required:
        - user
        - hero
        - dDayCards
        - subjects
        - recentActivities
      properties:
        user:
          $ref: '#/components/schemas/HomeUserSummary'
        hero:
          $ref: '#/components/schemas/HomeHero'
        dDayCards:
          type: array
          items:
            $ref: '#/components/schemas/SubjectExamSchedule'
        subjects:
          type: array
          items:
            $ref: '#/components/schemas/SubjectSummary'
        recentActivities:
          type: array
          items:
            $ref: '#/components/schemas/RecentActivity'
    HomeUserSummary:
      type: object
      required:
        - nickname
        - dotoriBalance
        - xpTotal
        - currentLevel
      properties:
        nickname:
          type: string
          example: 도토리장인
        dotoriBalance:
          type: integer
          minimum: 0
          example: 12
        xpTotal:
          type: integer
          minimum: 0
          example: 360
        currentLevel:
          type: integer
          minimum: 1
          maximum: 10
          example: 3
        levelName:
          type: string
          example: 펜굴리는 다람쥐
    HomeHero:
      type: object
      properties:
        hasActiveQuiz:
          type: boolean
          example: true
        activeQuiz:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/RecentActivity'
    RecentActivityListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiSuccessBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RecentActivityPage'
    RecentActivityPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/RecentActivity'
    RecentActivity:
      type: object
      required:
        - activityId
        - activityType
        - title
        - subjectId
        - subjectName
        - createdAt
      properties:
        activityId:
          type: string
          example: 018f8c2e-8123-72ab-96a1-f7c49c1d1111
        activityType:
          $ref: '#/components/schemas/RecentActivityType'
        quizSessionId:
          type: string
          format: uuid
          nullable: true
          example: 018f8c2e-8123-72ab-96a1-f7c49c1d1111
        playSessionId:
          type: string
          nullable: true
          example: android-play-session-uuid
        resultId:
          type: string
          format: uuid
          nullable: true
          description: 퀴즈 결과 publicId. activityType이 quiz_completed인 경우 사용
          example: 018f8c2e-9999-7b11-b2d1-774f4c1c2222
        title:
          type: string
          example: 데이터베이스 객관식 10문제
        subjectId:
          type: string
          format: uuid
          example: 018f8c2e-5f73-7b6a-b9f0-3f55e7f7c901
        subjectName:
          type: string
          example: 데이터베이스
        status:
          type: string
          nullable: true
          example: completed
        progressPct:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
          example: 80
        scoreText:
          type: string
          nullable: true
          example: 8/10
        createdAt:
          type: string
          format: date-time
          example: '2026-05-19T07:30:00+09:00'
    RecentActivityType:
      type: string
      enum:
        - quiz_generating
        - quiz_ready
        - quiz_in_progress
        - quiz_completed
        - lecture_uploaded
      example: quiz_completed
...
    SubjectSummary:
      type: object
      required:
        - id
        - name
        - purpose
        - chapterCount
        - partCount
      properties:
        id:
          type: string
          format: uuid
          example: 018f8c2e-5f73-7b6a-b9f0-3f55e7f7c901
        name:
          type: string
          minLength: 1
          maxLength: 30
          example: 데이터베이스
        purpose:
          $ref: '#/components/schemas/SubjectPurpose'
        chapterCount:
          type: integer
          minimum: 0
          example: 3
        partCount:
          type: integer
          minimum: 0
          example: 12
        lastActivityAt:
          type: string
          format: date-time
          nullable: true
          example: '2026-05-19T07:30:00+09:00'
        examSchedule:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/SubjectExamSchedule'
...
    SubjectPurpose:
      type: string
      enum:
        - exam
        - review
        - other
      example: exam
...
    SubjectExamSchedule:
      type: object
      required:
        - id
        - subjectId
        - examName
        - examDate
      properties:
        id:
          type: string
          format: uuid
          description: 시험 일정 publicId
          example: 018f8c2e-1234-7abc-91c1-000000000001
        subjectId:
          type: string
          format: uuid
          example: 018f8c2e-5f73-7b6a-b9f0-3f55e7f7c901
        examName:
          type: string
          example: 중간고사
        examDate:
          type: string
          format: date
          example: '2026-06-30'
        dDay:
          type: integer
          nullable: true
          description: 오늘 기준 남은 일수. 시험일이 지난 경우 null
          example: 3
...
    ApiSuccessBase:
      type: object
      required:
        - success
        - code
        - message
        - data
      properties:
        success:
          type: boolean
          example: true
        code:
          type: string
          example: COMMON_SUCCESS
        message:
          type: string
          example: 요청이 성공했습니다.
        data:
          nullable: true
    PageMeta:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
        - hasNext
      properties:
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 10
        totalElements:
          type: integer
          format: int64
          example: 42
        totalPages:
          type: integer
          example: 5
        hasNext:
          type: boolean
          example: true